@@ -33,55 +33,25 @@
  * @codingStandardsIgnoreStart
  */
 
-use Mollie\Api\Types\PaymentMethod;
+use _PhpScoper5ea00cc67502b\Mollie\Api\MollieApiClient;
+use _PhpScoper5ea00cc67502b\Mollie\Api\Types\PaymentMethod;
 
 class MolliePayScreenModuleFrontController extends ModuleFrontController
 {
 
     public function postProcess()
     {
-        $validateUrl = Context::getContext()->link->getModuleLink(
-            'mollie',
-            'payment',
-            array('method' => PaymentMethod::CREDITCARD, 'rand' => time()),
-            true
-        );
-
-        $cartId = $this->context->cart->id;
-        $secureKey = $this->context->customer->secure_key;
         $method = Tools::getValue('method');
         $cardToken = Tools::getValue('mollieCardToken' . $method);
 
-        $orderId = Order::getOrderByCartId($this->context->cart->id);
-        $mollie = new \Mollie\Api\MollieApiClient();
-        $mollie->setApiKey(Configuration::get(mollie::MOLLIE_API_KEY));
-
-
-        $orderLink = $this->context->link->getPageLink(
-            'order-confirmation',
-            true,
-            null,
-            [
-                'id_cart' => $cartId,
-                'id_module' => $this->module->name,
-                'id_order' => $orderId,
-                'key' => $secureKey,
-            ]
+        $validateUrl = Context::getContext()->link->getModuleLink(
+            'mollie',
+            'payment',
+            array('method' => PaymentMethod::CREDITCARD, 'rand' => time(), 'cardToken' => $cardToken),
+            true
         );
 
-        $payment = $mollie->payments->create([
-            "method" => "creditcard",
-            "amount" => [
-                "currency" => $this->context->currency->iso_code,
-                "value" => number_format($this->context->cart->getOrderTotal(),2)
-            ],
-            "description" => "Order #1",
-            "redirectUrl" => $validateUrl,
-            "webhookUrl" => "https://webshop.example.org/payments/webhook/",
-            "cardToken" => $cardToken,
-        ]);
-
-        Tools::redirect($payment->redirectUrl);
+        Tools::redirect($validateUrl);
     }
 
     public function initContent()
@@ -99,7 +69,7 @@ class MolliePayScreenModuleFrontController extends ModuleFrontController
     public function setMedia()
     {
         Media::addJsDef([
-            'profileId' => Configuration::get(Mollie::MOLLIE_PROFILE_ID),
+            'profileId' => Configuration::get(Mollie\Config\Config::MOLLIE_PROFILE_ID),
         ]);
         $this->addJS("{$this->module->getPathUri()}views/js/front/mollie_iframe.js");
         $this->addCSS("{$this->module->getPathUri()}views/css/mollie_iframe.css");
